+++ test/socketio/client_socket.io.test.js
@@ -1,0 +1,42 @@
+"use strict";
+
+const cds = require("@sap/cds");
+const auth = require("../_env/util/auth");
+
+cds.test(__dirname + "/../_env");
+
+cds.env.websocket.kind = "socket.io";
+
+describe("Client", () => {
+  let client;
+
+  beforeAll(async () => {
+    const port = cds.app.server.address().port;
+    client = await cds.connect.to("socket.io-client", {
+      url: `http://localhost:${port}/ws/chat`,
+      headers: {
+        authorization: auth.alice,
+      },
+    });
+  });
+
+  afterAll(async () => {
+    client.disconnect();
+    cds.ws.close();
+  });
+
+  test("Chat message", async () => {
+    const received = new Promise((resolve) => {
+      client.on("received", (message) => {
+        resolve(message);
+      });
+    });
+    client.enterContext("chat");
+    client.enterContext("test");
+    client.exitContext("test");
+    client.resetContexts();
+    client.emit("message", { text: "test" });
+    const result = await received;
+    expect(result.data).toEqual({ text: "test", user: "alice" });
+  });
+});