+++ test/ws/redis_ws.test.js
@@ -93,6 +93,9 @@ describe("Redis", () => {
           text: "test",
           user: "alice",
         },
+        headers: {
+          header: "value",
+        },
         event: "received",
       },
       {